added unit test

+++ TaskManagment/infrastructure/repositories/TaskRepository.cs
@@ -11,16 +11,50 @@ namespace infrastructure.repositories
 {
     public class TaskRepository : ITaskRepository
     {
-        private readonly NpgsqlDataSource _dataSource;
-
+        private readonly NpgsqlDataSource? _dataSource;
+        private readonly List<TaskModel> mockTasks;
         public TaskRepository(NpgsqlDataSource dataSource)
         {
-            _dataSource = dataSource;
+            if(dataSource != null)
+            {
+                _dataSource = dataSource;
+            }
+
+            mockTasks = new List<TaskModel>()
+            {
+                new TaskModel 
+                {
+                    Id = 1,
+                    TaskName = "MockTask1",
+                    TaskDescription = "This is a Mock Task",
+                    TaskStatus = false
+                },
+                new TaskModel 
+                {
+                    Id = 2,
+                    TaskName = "MockTask2",
+                    TaskDescription = "This is a Mock Task again",
+                    TaskStatus = true
+                },
+                new TaskModel
+                {
+                    Id = 3,
+                    TaskName = "MockTask3",
+                    TaskDescription = "This is a Mock Task haha",
+                    TaskStatus = true
+                }
+            };
+
         }
 
 
         public IEnumerable<TaskModel> GetAllTasks()
         {
+            if(_dataSource == null)
+            {
+                return mockTasks;
+            }
+
             using var conn = _dataSource.OpenConnection();
 
             return conn.Query<TaskModel>(@$"SELECT
@@ -39,6 +73,11 @@ namespace infrastructure.repositories
         }
         public TaskModel CreateTask(TaskModel task)
         {
+            if(_dataSource == null)
+            {
+                return task;
+            }
+            
             using var conn = _dataSource.OpenConnection();
 
             return conn.QueryFirst<TaskModel>($@"INSERT INTO public.tasks
@@ -53,9 +92,15 @@ namespace infrastructure.repositories
         }
         public void DeleteTask(int id)
         {
-            using var conn = _dataSource.OpenConnection();
-
-            conn.Execute($@"DELETE FROM public.tasks WHERE id=@id;", new {id});
+            if(_dataSource == null)
+            {
+                mockTasks.RemoveAll(task => task.Id == id);
+            }
+            else
+            {
+                using var conn = _dataSource.OpenConnection();
+                conn.Execute($@"DELETE FROM public.tasks WHERE id=@id;", new {id});
+            }
         }
     }
 }
